+++ adapter/yandex/src/main/java/org/bidon/yandex/YandexParameters.kt
@@ -1,0 +1,5 @@
+package org.bidon.yandex
+
+import org.bidon.sdk.adapter.AdapterParameters
+
+internal class YandexParameters : AdapterParameters